+++ tests/test_ticktock.py
@@ -179,3 +179,49 @@ def test_print(capfd: pytest.CaptureFixture):
             conditions[3] = True
 
     assert all(conditions)
+
+def test_exit_in_nested():
+    tt = TickTock()
+    with pytest.raises(ZeroDivisionError):
+        tt.profile("adfsadlæfj")
+        0 / 0
+        tt.end_profile()
+    # with tt.profile construct was not used, so expect an unclosed profile
+    assert len(tt._profile_stack) == 1
+
+    tt = TickTock()
+    with pytest.raises(ZeroDivisionError):
+        with tt.profile("sada"):
+            tt.profile("adfsadlæfj")
+            0 / 0
+            tt.end_profile()
+    # with tt.profile construct was used, so expect no unclosed profiles
+    assert len(tt._profile_stack) == 0
+
+    tt = TickTock()
+    with pytest.raises(ZeroDivisionError):
+        with tt.profile("adsfad"):
+            tt.profile("asdasd")
+            with tt.profile("sada"):
+                tt.profile("adfsadlæfj")
+                0 / 0
+                tt.end_profile()
+            tt.end_profile()
+    # with tt.profile construct was used, so expect no unclosed profiles
+    assert len(tt._profile_stack) == 0
+
+    tt = TickTock()
+    with pytest.raises(ZeroDivisionError):
+        tt.profile("aslkd")
+        tt.profile("alksjdasd")
+        with tt.profile("adsfad"):
+            tt.profile("asdasd")
+            with tt.profile("sada"):
+                tt.profile("adfsadlæfj")
+                0 / 0
+                tt.end_profile()
+            tt.end_profile()
+        tt.end_profile()
+        tt.end_profile()
+    # Two outermost did not using with construct, so expect two unclosed profiles
+    assert len(tt._profile_stack) == 2