+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support lists in moving average functions
 - Changed DataStorage to have default name based on class name
 - Non json-serializable data in DataStorage instances are now saved to a single pickle file
+- Prevent NameError being raised if an exception was raised inside nested profiling without the use of `with TT.profile`
 
 ## 2.0.0 - Breaking changes